fixed a regression in attribution test
