+++ pytest-selenium/pages/content.py
@@ -9,7 +9,7 @@ import random
 from pages.base import Page
 from regions.base import Region
 from regions.toc import TableOfContents
-from utils.utilities import WaitForTitleChange
+from utils.utility import WaitForTitleChange
 
 
 class Content(Page, WaitForTitleChange):
